Backend/requirements.txt
fastapi==0.135.1
google-auth==2.49.1
google-genai==1.68.0
httpx==0.28.1
pillow==12.1.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pytest==9.0.2
pytest-asyncio==1.3.0
python-jose==3.5.0
cryptography==42.0.0
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
uvicorn==0.42.0
